updated tests & doc strings
