+++ .github/workflows/test.yml
@@ -5,7 +5,7 @@ name: Testing
 
 on:
   push:
-    branches: [ "main" ]
+    branches: [ "main", "testing_framework" ]
   pull_request:
     branches: [ "main" ]
