+++ tests/test_OdmrLogic.py
@@ -26,6 +26,11 @@ import math
 import numpy as np
 import coverage
 import pytest
+import multiprocessing
+import rpyc
+from PySide2 import QtWidgets
+from PySide2.QtCore import QTimer
+from qudi.core import application
 from qudi.util.network import netobtain
 
 MODULE = 'odmr_logic'
@@ -33,6 +38,51 @@ BASE = 'logic'
 CHANNELS = ('APD counts', 'Photodiode')
 FIT_MODEL = 'Gaussian Dip'
 TOLERANCE = 10 # tolerance for signal data range
+CONFIG = os.path.join(os.getcwd(),'tests/test.cfg')
+
+
+def run_qudi(timeout=100000):
+    """
+    Runs a Qudi instance with a timer.
+
+    Parameters
+    ----------
+    timeout : int, optional
+        timeout for the Qudi session in milliseconds, by default 100000.
+    """
+    app_cls = QtWidgets.QApplication
+    app = app_cls.instance()
+    if app is None:
+        app = app_cls()
+    qudi_instance = application.Qudi.instance()
+    if qudi_instance is None:
+        qudi_instance = application.Qudi(config_file=CONFIG)
+    QTimer.singleShot(timeout, qudi_instance.quit)
+    qudi_instance.run()
+
+
+@pytest.fixture(scope='module')
+def start_qudi_process():
+    """
+    Fixture that starts the Qudi process and ensures it's running before returning.
+    """
+    qudi_process = multiprocessing.Process(target=run_qudi)
+    qudi_process.start()
+    yield
+    qudi_process.join(timeout=10)
+    if qudi_process.is_alive():
+        qudi_process.terminate()
+
+@pytest.fixture(scope='module')
+def remote_instance(start_qudi_process):
+    """
+    Fixture that connects to the running Qudi ipython kernel through rpyc client and returns the client instance.
+    """
+    time.sleep(5)
+    conn = rpyc.connect("localhost", 18861, config={'sync_request_timeout': 60})
+    root = conn.root
+    qudi_instance = root._qudi
+    return qudi_instance
 
 
 def get_scanner(module):
@@ -145,7 +195,7 @@ def coverage_for_each_test(request):
     print(f"Coverage report saved to {test_dir}")
 
 
-def test_start_odmr_scan(module, qtbot):
+def test_start_odmr_scan(module):
     """
     Tests if the scan parameters are correctly generated and if the signal data is generated for the given runtime
     with appropriate values.
