+++ tests/test_modify_status_var_example.py
@@ -20,18 +20,81 @@ You should have received a copy of the GNU Lesser General Public License along w
 If not, see <https://www.gnu.org/licenses/>.
 """
 
+import os
+import time
+import multiprocessing
+import rpyc
 import pytest
+from PySide2 import QtWidgets
+from PySide2.QtCore import QTimer
+from qudi.core import application
 from qudi.util.yaml import yaml_load, yaml_dump
 from qudi.util.paths import get_module_app_data_path
-
 
 LOGIC_MODULE = 'odmr_logic'
 GUI_MODULE = 'odmr_gui'
 STATUS_VAR = 'run_time'
 VALUE = 10
+CONFIG = os.path.join(os.getcwd(),'tests/test.cfg')
+
+
+def run_qudi(timeout=50000):
+    """
+    Runs a Qudi instance with a timer.
+
+    Parameters
+    ----------
+    timeout : int, optional
+        timeout for the Qudi session in milliseconds, by default 50000.
+    """
+    app_cls = QtWidgets.QApplication
+    app = app_cls.instance()
+    if app is None:
+        app = app_cls()
+    qudi_instance = application.Qudi.instance()
+    if qudi_instance is None:
+        qudi_instance = application.Qudi(config_file=CONFIG)
+    QTimer.singleShot(timeout, qudi_instance.quit)
+    qudi_instance.run()
+
+
+@pytest.fixture(scope='module')
+def start_qudi_process():
+    """
+    Fixture that starts the Qudi process and ensures it's running before returning.
+    """
+    qudi_process = multiprocessing.Process(target=run_qudi)
+    qudi_process.start()
+    yield
+    qudi_process.join(timeout=10)
+    if qudi_process.is_alive():
+        qudi_process.terminate()
+
+@pytest.fixture(scope='module')
+def remote_instance(start_qudi_process):
+    """
+    Fixture that connects to the running Qudi ipython kernel through rpyc client and returns the client instance.
+    """
+    time.sleep(5)
+    conn = rpyc.connect("localhost", 18861, config={'sync_request_timeout': 60})
+    root = conn.root
+    qudi_instance = root._qudi
+    return qudi_instance
+
+@pytest.fixture(scope='module')
+def logic_instance(remote_instance):
+    """ 
+    This fixture returns Odmr logic instance
+    """
+    module_manager = remote_instance.module_manager
+    module_manager.activate_module(GUI_MODULE)
+    logic_instance = module_manager._modules[LOGIC_MODULE].instance
+    return logic_instance
+
 
 def get_status_var_file(instance):
-    """This function returns the path for status variable file 
+    """
+    This function returns the path for status variable file 
 
     Parameters
     ----------
@@ -49,7 +112,8 @@ def get_status_var_file(instance):
     return file_path
 
 def load_status_var(file_path):    
-    """This function returns the loaded status variable from the file
+    """
+    This function returns the loaded status variable from the file
 
     Parameters
     ----------
@@ -69,7 +133,8 @@ def load_status_var(file_path):
     return variables
 
 def modify_status_var(status_vars, var, value):
-    """Setting status variable
+    """
+    Setting status variable
 
     Parameters
     ----------
@@ -89,7 +154,8 @@ def modify_status_var(status_vars, var, value):
     return status_vars
 
 def dump_status_variables(vars, file_path):
-    """Dump updated status variable to yaml
+    """
+    Dump updated status variable to yaml
 
     Parameters
     ----------
@@ -104,18 +170,9 @@ def dump_status_variables(vars, file_path):
         print("Failed to save status variables:", e)
 
 
-
-@pytest.fixture(scope='module')
-def logic_instance(remote_instance):
-    """This fixture returns Odmr logic instance"""
-    module_manager = remote_instance.module_manager
-    module_manager.activate_module(GUI_MODULE)
-    logic_instance = module_manager._modules[LOGIC_MODULE].instance
-    return logic_instance
-
-
 def test_status_vars(qudi_instance, qt_app):
-    """ Modifying a specific saved status variable for a specific module
+    """ 
+    Modifying a specific saved status variable for a specific module
 
     Parameters
     ----------
@@ -139,7 +196,8 @@ def test_status_vars(qudi_instance, qt_app):
     dump_status_variables(modified_vars, status_var_file_path)
 
 def test_status_vars_changed(logic_instance, qudi_instance):
-    """Test whether the status variable has changed
+    """
+    Test whether the status variable has changed
     
     Parameters
     ----------
